feature: add CMake build scripts
CMake allows for better projet interopability.
It is now possible to include this library in a CMake
project by calling add_subdirectory() and linking
the c4ev3::ev3api library.

It is forbidden to run build in the source directory,
because that would overwrite the hand-written Makefile.

Initially, only building libev3api.a is supported.

Additionally, three toolchain files are added:
 - c4ev3.cmake         - expects C4EV3 OSELAS toolchain on PATH.
 - codesoucery.cmake   - expects CS arm-2009q1-203 toolchain on PATH.
 - generic-armel.cmake - expects arm-linux-gnueabi-* on PATH.