feature: build Doxygen docs with CMake
It is now possible to build Doxygen documentation
with CMake. For this, the "docs" target is provided.
It will generate documentation in the <build>/docs
subdirectory.

The Doxyfile is now supplemented with a template named
Doxyfile.in. This file is processed by CMake, which
fills in input/output paths and project
version/name/description.