rename 'API' folder to 'src' and move Makefile up
The 'API' folder contains both source files of the library and headers/constants
from lms2012. The number of external headers is going to increase with the
future implementation of bluetooth messaging. To keep a clear distinction
between source and external files, we renamed the 'API' folder to 'src'. New
external files will be placed outside the src folder.

The Makefile has been moved outside src, to be able to also compile external
files.